pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiohttp>=3.11.18",
    "google-genai>=1.24.0",
    "python-dotenv>=1.1.1",
]
